User now can order single products by clicking on 'order' button

+++ orders/views.py
@@ -8,13 +8,13 @@ from django.db.models import Prefetch
 from django.views.generic import FormView, TemplateView
 from django.views import View
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.core.cache import cache
 
 from orders.forms import OrderForm
-from orders.utils import get_order_items
 from orders.models import Order, OrderProduct
 from users.models import User
+from products.models import Product
 from global_utils.mixins import CacheMixin
+from orders.utils import get_order_items
 
 
 class MakeOrderView(LoginRequiredMixin, FormView):
@@ -30,18 +30,22 @@ class MakeOrderView(LoginRequiredMixin, FormView):
         initial['phone_number'] = self.request.user.phone_number
         return initial
     
+    def dispatch(self, request, *args, **kwargs):
+        self.product_id = kwargs.get('product_id', None)
+        return super().dispatch(request, *args, **kwargs)
+
     def form_valid(self, form):
         try:
             with transaction.atomic():
                 user = self.request.user
-                user_order = get_order_items(request=self.request)
                 phone_number = form.cleaned_data['phone_number']
                 first_name = form.cleaned_data['first_name']
                 last_name = form.cleaned_data['last_name']
                 user_db = User.objects.filter(id=self.request.user.id)
+                product_id = self.product_id
 
-
-                if user_order.exists():
+                if product_id:
+                    user_order = Product.objects.get(id=product_id)
                     order = Order.objects.create(
                         user=user,
                         phone_number=phone_number,
@@ -49,20 +53,33 @@ class MakeOrderView(LoginRequiredMixin, FormView):
                         delivery_address=form.cleaned_data['delivery_address'],
                         payment_on_get=form.cleaned_data['payment_on_get'],
                     )
+                    OrderProduct.objects.create(order=order, product=user_order, name=user_order.name, price=user_order.price, quantity=1)
+                else:
+                    user_order = get_order_items(request=self.request)
 
-                    for prod in user_order:
-                        product = prod.product
-                        name = prod.product.name
-                        price = prod.product.price
-                        quantity = prod.quantity
-
-                        OrderProduct.objects.create(
-                            order=order,
-                            product=product,
-                            name=name,
-                            price=price,
-                            quantity=quantity,
+                    if user_order.exists():
+                        order = Order.objects.create(
+                            user=user,
+                            phone_number=phone_number,
+                            requires_delivery=form.cleaned_data['requires_delivery'],
+                            delivery_address=form.cleaned_data['delivery_address'],
+                            payment_on_get=form.cleaned_data['payment_on_get'],
                         )
+                        for prod in user_order:
+                                product = prod.product
+                                name = prod.product.name
+                                price = prod.product.price
+                                quantity = prod.quantity
+
+                                OrderProduct.objects.create(
+                                    order=order,
+                                    product=product,
+                                    name=name,
+                                    price=price,
+                                    quantity=quantity,
+                                )
+                        
+                        user_order.delete()
                 
                 if user.phone_number != phone_number:
                     user_db.update(phone_number=phone_number)
@@ -73,7 +90,8 @@ class MakeOrderView(LoginRequiredMixin, FormView):
                 if user.last_name != last_name:
                     user_db.update(last_name=last_name)
                 
-                user_order.delete()
+                if not product_id:
+                    user_order.delete()
 
                 return redirect('user:account')
